+++ bin/jpulse-install.sh
@@ -6,8 +6,8 @@
  #                  - Run as root: sudo npx jpulse install
  #                  - For Red Hat Enterprise Linux ecosystem
  # @file            bin/jpulse-install.sh
- # @version         1.1.6
- # @release         2025-11-14
+ # @version         1.1.7
+ # @release         2025-11-18
  # @repository      https://github.com/jpulse-net/jpulse-framework
  # @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  # @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/