+++ webapp/tests/helpers/websocket-test-utils.js
@@ -3,8 +3,8 @@
  * @tagline         WebSocket test utilities
  * @description     Mock WebSocket classes and helper functions for testing
  * @file            webapp/tests/helpers/websocket-test-utils.js
- * @version         1.1.6
- * @release         2025-11-14
+ * @version         1.1.7
+ * @release         2025-11-18
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/