+++ webapp/tests/unit/controller/markdown-ignore.test.js
@@ -3,8 +3,8 @@
  * @tagline         Unit tests for markdown controller ignore functionality
  * @description     Tests for .jpulse-ignore pattern matching and filtering
  * @file            webapp/tests/unit/controller/markdown-ignore.test.js
- * @version         1.1.6
- * @release         2025-11-14
+ * @version         1.1.7
+ * @release         2025-11-18
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/